fix: String methods for some types
