+++ cosmos-sdk/types/events.go
@@ -1,0 +1,10 @@
+package types
+
+import fmt "fmt"
+
+// StringAttributes defines a slice of StringEvents objects.
+type StringEvents []StringEvent
+
+func (a Attribute) String() string {
+	return fmt.Sprintf("%s: %s", a.Key, a.Value)
+}
